am62x: SOC: Fix Misra C Rule 10.8 issues
MISRA.ETYPE.COMP.CAST.EXPL.DIFFERENT.2012 Rule 10.8
The value of a composite expression shall not be cast to
a different essential type category or a wider essential type.

Fixes: SITSW-7018

+++ source/drivers/soc/am62x/soc.c
@@ -779,7 +779,7 @@ int32_t SOC_setPSCState(uint32_t instNum, uint32_t domainNum, uint32_t moduleNum
             /* Start the state transition */
             uint32_t pwrDmnGrp = domainNum >> 5U;
             uint32_t pwrDmnNumInGrp = domainNum & 0x1FU;
-            CSL_REG32_WR (baseAddr + CSL_PSC_PTCMD(pwrDmnGrp), 1 << pwrDmnNumInGrp);
+            CSL_REG32_WR ((uint32_t)baseAddr + CSL_PSC_PTCMD(pwrDmnGrp), 1U << pwrDmnNumInGrp);
 
             do {
                 pdTransStatus = CSL_FEXTR( baseAddr + CSL_PSC_PTSTAT(pwrDmnGrp), \
